requirements.txt
Flask==3.1.2
numpy==1.24.3
scikit-image==0.21.0
scipy==1.11.1
svgwrite==1.4.3
Pillow==10.0.0
networkx==3.1
